refactor: group configuration into a single package
