+++ app/src/main/kotlin/com/example/common/configs/OpenApiConfiguration.kt
@@ -1,4 +1,4 @@
-package com.example.common
+package com.example.common.configs
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition
 import io.swagger.v3.oas.annotations.info.Info
