+++ app/src/main/kotlin/com/example/common/configs/WebConfiguration.kt
@@ -1,4 +1,4 @@
-package com.example.common.web
+package com.example.common.configs
 
 import java.time.Duration
 import org.springframework.beans.factory.annotation.Value
