+++ app/src/main/kotlin/com/example/users/UserRepository.kt
@@ -14,7 +14,12 @@ interface UserRepository :
     @Query("select u from DomainUser u where u.username = :login or u.email = :login")
     fun findByLogin(login: String): DomainUser?
 
-    @Cacheable(cacheNames = [DEFAULT_CACHE], key = "#username", condition = "#username == #email")
+    @Cacheable(
+        cacheNames = [DEFAULT_CACHE],
+        key = "#username",
+        condition = "#username == #email",
+        unless = "#result == null",
+    )
     fun findByUsernameOrEmailIgnoreCase(username: String, email: String): DomainUser?
 
     fun findByUsernameIgnoreCase(username: String): DomainUser?
