+++ app/src/test/kotlin/com/example/common/JPATestConfiguration.kt
@@ -1,6 +1,6 @@
 package com.example.common
 
-import com.example.common.persistence.JpaConfiguration
+import com.example.common.configs.JPAConfiguration
 import java.util.Optional
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.context.annotation.Bean
@@ -10,7 +10,7 @@ import org.springframework.data.domain.AuditorAware
 const val DEFAULT_TEST_AUDITOR = "test-auditor"
 
 @TestConfiguration
-@Import(JpaConfiguration::class)
+@Import(JPAConfiguration::class)
 class JPATestConfiguration {
 
     @Bean
