+++ app/src/test/kotlin/com/example/users/UsersRepositoryTest.kt
@@ -2,7 +2,7 @@ package com.example.users
 
 import com.example.common.DEFAULT_TEST_AUDITOR
 import com.example.common.JPATestConfiguration
-import com.example.common.security.AuthorityConstants
+import com.example.common.configs.AuthorityConstants
 import java.time.Instant
 import org.assertj.core.api.Java6Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
